Document Scala 3 limitations by running tests with different outputs
These expected outputs intentionally retain wildcard imports because the current
Scala 3 SemanticDB owner model does not produce the explicit expansions expected
by the Scala 2 fixtures.

- `ExpandWildcard.scala`: nested module members do not expose the owner
  information required for expansion.
- `ExpandWildcardImplicit.scala`: synthetic implicit references are not exposed
  under the wildcard owner in the form required for expansion.
- `ExpandWildcardInherited.scala`: inherited members use a different owner
  chain, so the rule cannot match both references.
- `ExpandWildcardInstanceCall.scala`: the instance-call member is not modeled
  through the qualifier type in the form required for expansion.
- `ExpandWildcardPackageObject.scala`: package-object owners are represented
  differently, so these members cannot be modeled for expansion.
- `ExpandWildcardQuotedIdent.scala`: quoted identifiers expose different
  SemanticDB names and owners.
- `ExpandWildcardWithExplicit.scala`: the unqualified `b` is not associated
  with the wildcard beside explicit `a`.
