+++ scalafix-tests/input/src/main/scala/test/organizeImports/ExpandWildcardImplicit.scala
@@ -4,14 +4,14 @@ OrganizeImports {
   expandWildcardImportThreshold = 5
   groupedImports = Keep
   removeUnused = false
-  targetDialect = Scala2
+  targetDialect = Auto
 }
  */
 package test.organizeImports
 
-import test.organizeImports.ExpandInheritance.Sub._
+import test.organizeImports.Implicits.a._
 
-object ExpandWildcardInherited {
-  val a = own
-  val b = inherited
+object ExpandWildcardImplicit {
+  def needsInt(implicit i: Int): Int = i
+  val x: Int = needsInt
 }
