+++ scalafix-tests/input/src/main/scala/test/organizeImports/ExpandWildcardQuotedIdent.scala
@@ -4,14 +4,14 @@ OrganizeImports {
   expandWildcardImportThreshold = 5
   groupedImports = Keep
   removeUnused = false
-  targetDialect = Scala2
+  targetDialect = Auto
 }
  */
 package test.organizeImports
 
-import test.organizeImports.MergeImports.Wildcard1.{a, _}
+import test.organizeImports.QuotedIdent._
 
-object ExpandWildcardWithExplicit {
-  val x1 = a
-  val x2 = b
+object ExpandWildcardQuotedIdent {
+  val x = `macro`
+  val y = `a.b`
 }
