+++ scalafix-tests/input/src/main/scala/test/organizeImports/ExpandWildcardSamePackage.scala
@@ -4,7 +4,7 @@ OrganizeImports {
   expandWildcardImportThreshold = 5
   groupedImports = Keep
   removeUnused = false
-  targetDialect = Scala2
+  targetDialect = Auto
 }
  */
 package test.organizeImports.samepkg
