+++ scalafix-tests/output/src/main/scala-3/test/organizeImports/ExpandWildcard.scala
@@ -1,0 +1,8 @@
+package test.organizeImports
+
+import test.organizeImports.MergeImports.Wildcard1.*
+
+object ExpandWildcard {
+  val x1 = a
+  val x2 = b
+}
