+++ scalafix-tests/output/src/main/scala-3/test/organizeImports/ExpandWildcardImplicit.scala
@@ -1,0 +1,8 @@
+package test.organizeImports
+
+import test.organizeImports.Implicits.a.*
+
+object ExpandWildcardImplicit {
+  def needsInt(implicit i: Int): Int = i
+  val x: Int = needsInt
+}
