+++ scalafix-tests/output/src/main/scala-3/test/organizeImports/ExpandWildcardInherited.scala
@@ -1,0 +1,8 @@
+package test.organizeImports
+
+import test.organizeImports.ExpandInheritance.Sub.*
+
+object ExpandWildcardInherited {
+  val a = own
+  val b = inherited
+}
