+++ scalafix-tests/output/src/main/scala-3/test/organizeImports/ExpandWildcardPackageObject.scala
@@ -1,0 +1,8 @@
+package test.organizeImports
+
+import test.organizeImports.pkgobj.*
+
+object ExpandWildcardPackageObject {
+  val a = directMember
+  val b = inheritedMember
+}
