+++ scalafix-tests/output/src/main/scala-3/test/organizeImports/ExpandWildcardQuotedIdent.scala
@@ -1,0 +1,8 @@
+package test.organizeImports
+
+import test.organizeImports.QuotedIdent.*
+
+object ExpandWildcardQuotedIdent {
+  val x = `macro`
+  val y = `a.b`
+}
