+++ scalafix-tests/output/src/main/scala-3/test/organizeImports/ExpandWildcardThreshold.scala
@@ -1,0 +1,9 @@
+package test.organizeImports
+
+import test.organizeImports.MergeImports.Wildcard1.*
+
+object ExpandWildcardThreshold {
+  val x1 = a
+  val x2 = b
+  val x3 = c
+}
