+++ scalafix-tests/output/src/main/scala-3/test/organizeImports/ExpandWildcardUnused.scala
@@ -1,0 +1,7 @@
+package test.organizeImports
+
+import test.organizeImports.MergeImports.Wildcard1.*
+
+object ExpandWildcardUnused {
+  val x = 1
+}
