+++ scalafix-tests/output/src/main/scala-3/test/organizeImports/ExpandWildcardWithUnimport.scala
@@ -1,0 +1,7 @@
+package test.organizeImports
+
+import test.organizeImports.MergeImports.Wildcard1.{a as _, *}
+
+object ExpandWildcardWithUnimport {
+  val x = b
+}
