+++ scalafix-tests/output/src/main/scala/test/organizeImports/ExpandWildcardSamePackage.scala
