+++ scalafix-tests/shared/src/main/scala/test/organizeImports/ExpandPkgObj.scala
