+++ scalafix-tests/shared/src/main/scala/test/organizeImports/ExpandSamePkg.scala
