+++ scalafix-tests/shared/src/main/scala/test/organizeImports/InstanceCall.scala
